gsdx ogl: fix previous commit * clean shader properly * use a 64 bits map to contain shader pointer
git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5631 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ plugins/GSdx/GSDeviceOGL.h
@@ -658,7 +658,7 @@ class GSDeviceOGL : public GSDevice
 	void SetupOM(OMDepthStencilSelector dssel, OMBlendSelector bsel, uint8 afix);
 
 #ifdef DISABLE_GL41_SSO
-	hash_map<uint32, GLuint > m_single_prog;
+	hash_map<uint64, GLuint > m_single_prog;
 	GLuint link_prog();
 #endif
